requirements.txt
streamlit
pymupdf
pandas
openpyxl
Pillow
